move credits route
